Lab3A completed (#4)
* initial check-in for lab3

fix build issue

* lab3A first imple

one client, concurrent client test cases work.
Basicaly logic done

* change to remember last leader

* clean up unneed code

* server to detect that it has lost leadership

handle the case in which a leader has called Start() for a Clerk's RPC,
but loses its leadership before the request is committed to the log. In
this case you should arrange for the Clerk to re-send the request to
other servers until it finds the new leader. One way to do this is for
the server to detect that it has lost leadership, by noticing that a
different request has appeared at the index returned by Start(), or that
Raft's term has changed.

* uniquely identify client operations part1

ensure that the key/value service executes each one just once.

* add uuid support for Get

also, fixed some race condition

* validation for term should move up

* lab3a done

refactor code

* update result